pyproject.toml
[tool.poetry]
name = "prevalidate"
version = "0.1.0"
description = ""
authors = ["Joey Dreijer <jdreijer@d3vzer0.com>"]
readme = "README.md"

[tool.poetry.scripts]
prevalidate = "prevalidate.main:app"

[tool.poetry.dependencies]
python = "<3.12,>=3.8.1"
pydantic = "^1.10.7"
pythonnet = "^3.0.1"
typer = "^0.9.0"
azure-identity = "^1.13.0"
azure-mgmt-loganalytics = "13.0.0b6"
pyyaml = "^6.0"
azure-monitor-query = "^1.2.0"
jinja2 = "^3.1.2"
pytest = "^7.3.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
